requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-magic==0.4.27
SQLAlchemy==2.0.27
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
